Add files via upload
Excited to kickstart my Data Science Internship at Prodigy InfoTech InfoTech!

For my first task, Task-01, I'll be working on data visualization. Specifically, I'll create a bar chart or histogram to analyze the distribution of a variable-like age or gender-in a dataset. This is a crucial step in understanding the underlying patterns and gaining insights.

To get started, I'll use a sample dataset, such as the World Bank's population data, which provides rich information for exploration.

Looking forward to applying my skills, learning new techniques, and making meaningful contributions during this journey!

#DataScience

#Internship Journey

#DataVisualization

#WorldBankData